Clean up examples to be in a solution folder and avoid unnecessary folder debt
