+++ examples/Common/Common.csproj
@@ -6,10 +6,11 @@
     <ImplicitUsings>enable</ImplicitUsings>
     <Nullable>enable</Nullable>
     <IsPackable>false</IsPackable>
+    <RootNamespace>Common</RootNamespace>
   </PropertyGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\..\..\src\Concordium.Sdk.csproj" />
+    <ProjectReference Include="..\..\src\Concordium.Sdk.csproj" />
   </ItemGroup>
 
   <ItemGroup>
