+++ examples/Common/Example.cs
