+++ examples/Common/ExampleOptions.cs
