+++ examples/Common/ExampleOptionsWithKeys.cs
