--- examples/Examples/Common/Concordium.Sdk.Examples.Common.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 25.0.1705.3
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Concordium.Sdk.Examples.Common", "Concordium.Sdk.Examples.Common.csproj", "{F2ECDF92-D479-4755-9FDE-73DE76AA5D26}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{F2ECDF92-D479-4755-9FDE-73DE76AA5D26}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{F2ECDF92-D479-4755-9FDE-73DE76AA5D26}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{F2ECDF92-D479-4755-9FDE-73DE76AA5D26}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{F2ECDF92-D479-4755-9FDE-73DE76AA5D26}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {461C702A-0E13-488E-AF98-CA6F4AE052D2}
-	EndGlobalSection
-EndGlobal
