+++ examples/GetAccountInfo/Program.cs
