+++ examples/GetBlockItemStatus/Program.cs
