+++ examples/GetBlockItemStatus/Transactions.GetBlockItemStatus.csproj
@@ -6,10 +6,11 @@
     <ImplicitUsings>enable</ImplicitUsings>
     <Nullable>enable</Nullable>
     <IsPackable>false</IsPackable>
+    <RootNamespace>Transactions.GetBlockItemStatus</RootNamespace>
   </PropertyGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\..\..\..\src\Concordium.Sdk.csproj" />
+    <ProjectReference Include="..\..\src\Concordium.Sdk.csproj" />
   </ItemGroup>
 
   <ItemGroup>
