+++ examples/GetFinalizedBlocks/Program.cs
