+++ examples/RegisterData/Program.cs
@@ -3,21 +3,21 @@ using Concordium.Sdk.Examples.Common;
 using Concordium.Sdk.Types;
 using Concordium.Sdk.Wallets;
 
-namespace Concordium.Sdk.Examples.Transactions.Transfer;
+namespace Transactions.RegisterData;
 
 /// <summary>
-/// Example demonstrating how to submit a transfer transaction.
+/// Example demonstrating how to submit a register data transaction.
 ///
 /// The example assumes you have your account key information stored
 /// in the Concordium browser wallet key export format, and that a path
 /// pointing to it is supplied to it from the command line.
 ///
-/// Cfr. <see cref="TransferTransactionExampleOptions"/> for more info
-/// on how to run the program, or refer to the help message.
+/// Cfr. <see cref="RegisterDataTransactionExampleOptions"/> for more
+/// info on how to run the program, or refer to the help message.
 /// </summary>
 internal class Program
 {
-    private static void SendTransferTransaction(TransferTransactionExampleOptions options)
+    private static void SendRegisterDataTransaction(RegisterDataTransactionExampleOptions options)
     {
         // Read the account keys from a file.
         var walletData = File.ReadAllText(options.WalletKeysFile);
@@ -30,10 +30,9 @@ internal class Program
             options.Timeout
         );
 
-        // Create the transfer transaction.
-        var amount = CcdAmount.FromCcd(options.Amount);
-        var receiver = AccountAddress.From(options.Receiver);
-        var transferPayload = new Sdk.Transactions.Transfer(amount, receiver);
+        // Encode a string as CBOR and use that as the data to register.
+        var data = OnChainData.FromTextEncodeAsCBOR(options.Data);
+        var transferPayload = new Concordium.Sdk.Transactions.RegisterData(data);
 
         // Prepare the transaction for signing.
         var sender = account.AccountAddress;
@@ -44,13 +43,13 @@ internal class Program
         // Sign the transaction using the account keys.
         var signedTransfer = preparedTransfer.Sign(account);
 
-        // Submit the transaction.
+        // Try to submit the transaction.
         var txHash = client.SendAccountTransaction(signedTransfer);
 
         // Print the transaction hash.
-        Console.WriteLine($"Successfully submitted transfer transaction with hash {txHash}");
+        Console.WriteLine($"Successfully submitted register data transaction with hash {txHash}");
     }
 
     private static void Main(string[] args) =>
-        Example.Run<TransferTransactionExampleOptions>(args, SendTransferTransaction);
+        Example.Run<RegisterDataTransactionExampleOptions>(args, SendRegisterDataTransaction);
 }
