+++ examples/RegisterData/Transactions.RegisterData.csproj
@@ -9,7 +9,6 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\..\Common\Concordium.Sdk.Examples.Common.csproj" />
-    <ProjectReference Include="..\..\..\..\src\Concordium.Sdk.csproj" />
+    <ProjectReference Include="..\Common\Common.csproj" />
   </ItemGroup>
 </Project>
