+++ examples/Transfer/Program.cs
@@ -3,21 +3,21 @@ using Concordium.Sdk.Examples.Common;
 using Concordium.Sdk.Types;
 using Concordium.Sdk.Wallets;
 
-namespace Concordium.Sdk.Examples.Transactions.RegisterData;
+namespace Transactions.Transfer;
 
 /// <summary>
-/// Example demonstrating how to submit a register data transaction.
+/// Example demonstrating how to submit a transfer transaction.
 ///
 /// The example assumes you have your account key information stored
 /// in the Concordium browser wallet key export format, and that a path
 /// pointing to it is supplied to it from the command line.
 ///
-/// Cfr. <see cref="RegisterDataTransactionExampleOptions"/> for more
-/// info on how to run the program, or refer to the help message.
+/// Cfr. <see cref="TransferTransactionExampleOptions"/> for more info
+/// on how to run the program, or refer to the help message.
 /// </summary>
 internal class Program
 {
-    private static void SendRegisterDataTransaction(RegisterDataTransactionExampleOptions options)
+    private static void SendTransferTransaction(TransferTransactionExampleOptions options)
     {
         // Read the account keys from a file.
         var walletData = File.ReadAllText(options.WalletKeysFile);
@@ -30,9 +30,10 @@ internal class Program
             options.Timeout
         );
 
-        // Encode a string as CBOR and use that as the data to register.
-        var data = OnChainData.FromTextEncodeAsCBOR(options.Data);
-        var transferPayload = new Sdk.Transactions.RegisterData(data);
+        // Create the transfer transaction.
+        var amount = CcdAmount.FromCcd(options.Amount);
+        var receiver = AccountAddress.From(options.Receiver);
+        var transferPayload = new Concordium.Sdk.Transactions.Transfer(amount, receiver);
 
         // Prepare the transaction for signing.
         var sender = account.AccountAddress;
@@ -43,13 +44,13 @@ internal class Program
         // Sign the transaction using the account keys.
         var signedTransfer = preparedTransfer.Sign(account);
 
-        // Try to submit the transaction.
+        // Submit the transaction.
         var txHash = client.SendAccountTransaction(signedTransfer);
 
         // Print the transaction hash.
-        Console.WriteLine($"Successfully submitted register data transaction with hash {txHash}");
+        Console.WriteLine($"Successfully submitted transfer transaction with hash {txHash}");
     }
 
     private static void Main(string[] args) =>
-        Example.Run<RegisterDataTransactionExampleOptions>(args, SendRegisterDataTransaction);
+        Example.Run<TransferTransactionExampleOptions>(args, SendTransferTransaction);
 }
