+++ examples/TransferWithMemo/TransferWithMemoTransactionExampleOptions.cs
@@ -2,7 +2,7 @@ using CommandLine;
 
 using Concordium.Sdk.Examples.Common;
 
-namespace Concordium.Sdk.Examples.Transactions.TransferWithMemo;
+namespace Transactions.TransferWithMemo;
 
 public class TransferWithMemoTransactionExampleOptions : ExampleOptionsWithKeys
 {
